+++ test/all_tests.dart
@@ -5,6 +5,7 @@ import 'memory_pool_test.dart' as mempool;
 import 'persistent_map_map_test.dart' as mapmap;
 import 'persistent_map_test.dart' as map;
 import 'persistent_string_map_test.dart' as stringmap;
+import 'regression_test.dart' as regression;
 
 void main() {
   mempool.main();
@@ -12,4 +13,5 @@ void main() {
   map.main();
   stringmap.main();
   mapmap.main();
+  regression.main();
 }