+++ test/persistent_map_test.dart
@@ -3,7 +3,7 @@ import 'package:libdbm/libdbm.dart';
 import 'package:test/test.dart';
 
 void main() {
-  var file = File('dummy.bin');
+  var file = File('dummy.persistent_map.bin');
   setUp(() async {
     if (file.existsSync()) {
       try {
@@ -49,6 +49,21 @@ void main() {
     expect(() => map.containsValue(null), throwsA(isA<AssertionError>()));
     expect(() => map.update('foo', (v) => 'bar', ifAbsent: null),
         throwsA(isA<AssertionError>()));
+    map.close();
+  });
+  test('putIfAbsent callback is lazy', () {
+    final map = PersistentMap.withStringValue(file, create: true);
+    map['foo'] = 'bar';
+
+    var calls = 0;
+    final value = map.putIfAbsent('foo', () {
+      calls += 1;
+      return 'baz';
+    });
+
+    expect(value, equals('bar'));
+    expect(calls, equals(0));
+    expect(map['foo'], equals('bar'));
     map.close();
   });
 }